Change Environment Parameter to State for Functions (#24)
Previously functions received &Environment. This is now changed to &State which
is a new proxy type that gives access to the internals of the current state of the
engine. This provides a more useful abstraction as interfacing with the current
execution becomes possible. This provides a similar level of flexibility as the
pass_context decorator in Jinja.

This also introduces the debug() function which spits out the current state in
stringified form.