Update minijinja cabi to account for usize changes
Hi! 

While trying to build minijinja-cabi for Wasm targets, I encountered an error due to a size mismatch on transmute operations, due to `usize` being 32 bits on those targets which breaks existing size assumptions, this attempts to fix this by doing the following:

* Add `const VALUE_SIZE: usize = size_of::<Value>();`
* Change `_opaque: [usize; 3],` to `_opaque: [u8; VALUE_SIZE],` in the `mj_value` struct.
* Change `transmute::<Value, [usize; 3]>(value)` to `transmute::<Value, [u8; VALUE_SIZE]>(value)` in the `impl From<Value> for mj_value` block.

Does this look like the correct approach?