Check that types passed to ViaDeserialize implement Serialize
This provides much better feedback to users who are trying to use a type
that does not implement `Deserialize` with `ViaDeserialize`. Previously,
the error would appear when trying to use a function or type that itself
uses `ViaDeserialize`, e.g. when calling `Environment::add_function` on
a function with a parameter using `ViaDeserialize` on a type that does
not implement `Deserialize`. For example:

```
error[E0277]: the trait bound `for<'a, 'b, 'c> fn(ViaDeserialize<&'a Page<'b>>, ViaDeserialize<&'c data::config::Config>) -> std::string::String {url_for}: Function<_, _>` is not satisfied
   --> src/templates/functions.rs:66:32
    |
66  |    env.add_function("url_for", url_for);
    |        ------------            ^^^^^^^ the trait `Function<_, _>` is not implemented for fn item `fn(ViaDeserialize<&Page<'b>>, ViaDeserialize<&Config>) -> String {url_for}`
    |        |
    |        required by a bound introduced by this call
    |
note: required by a bound in `Environment::<'source>::add_function`
   --> /Users/chris/.cargo/registry/src/index.crates.io-6f17d22bba15001f/minijinja-2.5.0/src/environment.rs:740:12
    |
736 |     pub fn add_function<N, F, Rv, Args>(&mut self, name: N, f: F)
    |            ------------ required by a bound in this associated function
...
740 |         F: functions::Function<Rv, Args>
    |            ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ required by this bound in `Environment::<'source>::add_function`

```

With this change in place, the error appears at the site where the user
tries to use the type incorrectly. For example:

```
error[E0277]: the trait bound `for<'de> NoDeserialize: Deserialize<'de>` is not satisfied
   --> minijinja/tests/test_value.rs:607:18
    |
607 |     let invalid: ViaDeserialize<NoDeserialize>;
    |                  ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ the trait `for<'de> Deserialize<'de>` is not implemented for `NoDeserialize`
    |
    = note: for local types consider adding `#[derive(serde::Deserialize)]` to your `NoDeserialize` type
    = note: for types from other crates check whether the crate offers a `serde` feature flag
    = help: the following other types implement trait `Deserialize<'de>`:
              &'a Path
              &'a [u8]
              &'a str
              ()
              (T,)
              (T0, T1)
              (T0, T1, T2)
              (T0, T1, T2, T3)
            and 152 others
note: required by a bound in `ViaDeserialize`
   --> /Users/chris/dev/mitsuhiko/minijinja/minijinja/src/value/deserialize.rs:113:30
    |
113 | pub struct ViaDeserialize<T: for<'de> serde::Deserialize<'de>>(pub T);
    |                              ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ required by this bound in `ViaDeserialize`
```

This is still far from the best possible error message, but it has two
major advantages over the previous situation:

1. It is local, that is, the error message appears at the point the user
   makes the mistake.
2. It is actually clear what the issue is: the type in question does not
   implement `Deserialize`.

Unfortunately, this does require introducing a higher-ranked trait bound
on the type parameter `T` for `ViaDeserialize`, and propagating it to
the impls for `ArgType`, `Deref`, and `DerefMut`. What is more, there is
today no *great* way to test this: the tests we would need to write are
tests that output compiler errors.

This is not, as far as I can tell, a breaking change: no type which
worked with `ViaDeserialize` before should work today, and vice versa;
the only effect *should* be changing where the error appears.