Fix the function reference in the src/tests docs
The documentation for the tests module referenced `is_defined`, but
explained it as "check if the value is indeed an odd number". Corrected
the example to use `is odd`, and the function name to `is_odd`.

Could have corrected the wording too, but the next example shows
even/odd, so correcting the test and the function name felt more
appropriate.

Signed-off-by: Gergely Nagy <me@gergo.csillger.hu>