Prepare for supporting two format styles
Add an enum capturing two options: printf-style, and str.format style

Move functions related to managing a pointer within input string into a new Cursor
type. Change parser functions to take a cursor as argument, use it to read chars from
input, and advance it as chars are consumed.

Add a Tokenizer type tokenizing input string into literal vs. replacement
fields. It'll invoke a style-specific parser to parse the field when a delimiter ('%'
or '{') is encountered.