Support general format ('g' and 'G'), and fix 'e' and default format
The general format rounds the floating point number to given precision `p`
significant digits, and then picks either fixed-point format or scientific notation,
based on the magnitude of the number. It first formats the number with `e` notation
and with `p - 1` precision (note that the `p` is number of significant digits in case
of general format, unlike `e` or `f` format, where it's number of digits after the
decimal point). If this scientific form has exponent, `exp`, within the range `-4 <=
exp < p`, then the fixed-point format using `f` presentation and `p-1-exp` precision
is picked; otherwise the scientific form is used.

Fix the `e` format to print the exponent in the same format as Python's string
formatter does: a sign is always printed, followed by minimum 2-width integer, padded
with `0` if necessary. Rust's fmt library doesn't follow the same format. So take its
output and format it according to Python's format to be compatible with jinja2.

If presentation type is missing in the spec, the general format `g` should be used
for the floating point numbers. So update the formatter accordingly.