Fix the formatting of floats when the spec is missing
If there's no spec provided in the input, prepare the `FormatSpec` token with default
values, and use that to format the given values, rather than falling back to
minijinja Value's Display trait. This fixes `{}` to format floats the same way as
`{:}` would do, i.e, using the format similar to `g`.